# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DevOps Console后端API文档",
        "title": "DevOps Console API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/instance/add": {
            "post": {
                "description": "添加新的Elasticsearch实例配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instance"
                ],
                "summary": "添加Elasticsearch实例",
                "parameters": [
                    {
                        "description": "实例信息",
                        "name": "instance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Instance"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    }
                }
            }
        },
        "/api/instance/delete": {
            "get": {
                "description": "根据ID删除指定的Elasticsearch实例",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instance"
                ],
                "summary": "删除Elasticsearch实例",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "实例ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "404": {
                        "description": "实例不存在",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    }
                }
            }
        },
        "/api/instance/get": {
            "get": {
                "description": "根据ID获取指定Elasticsearch实例的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instance"
                ],
                "summary": "获取Elasticsearch实例详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "实例ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "404": {
                        "description": "实例不存在",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    }
                }
            }
        },
        "/api/instance/instance-types": {
            "get": {
                "description": "获取所有可用的实例类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instance"
                ],
                "summary": "获取实例类型列表",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    }
                }
            }
        },
        "/api/instance/list": {
            "get": {
                "description": "获取Elasticsearch实例列表，支持分页和过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instance"
                ],
                "summary": "获取Elasticsearch实例列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "实例状态过滤",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "实例类型过滤",
                        "name": "type_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键词搜索",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    }
                }
            }
        },
        "/api/instance/test-connection": {
            "post": {
                "description": "测试指定实例的连接状态和响应时间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instance"
                ],
                "summary": "测试实例连接",
                "parameters": [
                    {
                        "description": "连接测试请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ConnectionTestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    }
                }
            }
        },
        "/api/instance/test-history": {
            "get": {
                "description": "获取指定实例的连接测试历史记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instance"
                ],
                "summary": "获取测试历史记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "实例ID",
                        "name": "instance_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页记录数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    }
                }
            }
        },
        "/api/instance/today-test-stats": {
            "get": {
                "description": "获取今日所有实例的连接测试统计信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instance"
                ],
                "summary": "获取今日测试统计",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    }
                }
            }
        },
        "/api/instance/update": {
            "post": {
                "description": "更新现有Elasticsearch实例的配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instance"
                ],
                "summary": "更新Elasticsearch实例",
                "parameters": [
                    {
                        "description": "实例信息",
                        "name": "instance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Instance"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/common.ReturnData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.ReturnData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.AuthConfig": {
            "type": "object",
            "properties": {
                "auth_type": {
                    "type": "string"
                },
                "config_key": {
                    "type": "string"
                },
                "config_value": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_encrypted": {
                    "type": "boolean"
                },
                "resource_id": {
                    "type": "integer"
                },
                "resource_name": {
                    "description": "资源名称",
                    "type": "string"
                },
                "resource_type": {
                    "description": "instance或cluster",
                    "type": "string"
                },
                "status": {
                    "description": "配置状态",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Instance": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "auth_configs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AuthConfig"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "https_enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "instance_type": {
                    "description": "关联",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.InstanceType"
                        }
                    ]
                },
                "instance_type_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "skip_ssl_verify": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.InstanceType": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instances": {
                    "description": "关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Instance"
                    }
                },
                "type_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "request.ConnectionTestRequest": {
            "type": "object",
            "required": [
                "instance_id"
            ],
            "properties": {
                "instance_id": {
                    "description": "实例ID",
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}